# app/temp_files/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Coffee point",
    "description": "This backend application is built on FastAPI and implements the full logic of cafe management.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/check-lists/start/": {
      "get": {
        "tags": [
          "check_lists"
        ],
        "summary": "Get Start Check List",
        "operationId": "get_start_check_list_check_lists_start__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckListOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/check-lists/end/": {
      "get": {
        "tags": [
          "check_lists"
        ],
        "summary": "Get End Check List",
        "operationId": "get_end_check_list_check_lists_end__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckListOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/check-lists/{check_list_id}/": {
      "put": {
        "tags": [
          "check_lists"
        ],
        "summary": "Update Start Check List",
        "operationId": "update_start_check_list_check_lists__check_list_id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "check_list_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Check List Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckListUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clients/{client_id}/": {
      "get": {
        "tags": [
          "clients"
        ],
        "summary": "Get Client",
        "operationId": "get_client_clients__client_id___get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Client Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "clients"
        ],
        "summary": "Update Client",
        "operationId": "update_client_clients__client_id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Client Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clients/": {
      "get": {
        "tags": [
          "clients"
        ],
        "summary": "Get Clients",
        "operationId": "get_clients_clients__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClientOut"
                  },
                  "type": "array",
                  "title": "Response Get Clients Clients  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "clients"
        ],
        "summary": "Create Client",
        "operationId": "create_client_clients__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/employees/": {
      "get": {
        "tags": [
          "employees"
        ],
        "summary": "Get Employees",
        "operationId": "get_employees_employees__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EmployeeOut"
                  },
                  "type": "array",
                  "title": "Response Get Employees Employees  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "employees"
        ],
        "summary": "Create Employee",
        "operationId": "create_employee_employees__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/employees/available/": {
      "get": {
        "tags": [
          "employees"
        ],
        "summary": "Get Available Employees",
        "operationId": "get_available_employees_employees_available__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EmployeeOut"
                  },
                  "type": "array",
                  "title": "Response Get Available Employees Employees Available  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/employees/{employee_id}/": {
      "put": {
        "tags": [
          "employees"
        ],
        "summary": "Update Employee",
        "operationId": "update_employee_employees__employee_id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Employee Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "employees"
        ],
        "summary": "Deactivate Employee",
        "operationId": "deactivate_employee_employees__employee_id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Employee Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/employee_shifts/": {
      "get": {
        "tags": [
          "employee_shifts"
        ],
        "summary": "Get Employee Shifts",
        "operationId": "get_employee_shifts_employee_shifts__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EmployeeShiftOut"
                  },
                  "type": "array",
                  "title": "Response Get Employee Shifts Employee Shifts  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "employee_shifts"
        ],
        "summary": "Create Employee Shift",
        "operationId": "create_employee_shift_employee_shifts__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeShiftCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeShiftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/employee_shifts/active-employee-shifts/": {
      "get": {
        "tags": [
          "employee_shifts"
        ],
        "summary": "Get Active Employee Shifts",
        "operationId": "get_active_employee_shifts_employee_shifts_active_employee_shifts__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EmployeeShiftWithEmployeeOut"
                  },
                  "type": "array",
                  "title": "Response Get Active Employee Shifts Employee Shifts Active Employee Shifts  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/employee_shifts/shift-end-update/{shift_id}/": {
      "put": {
        "tags": [
          "employee_shifts"
        ],
        "summary": "Update Employee Shift End",
        "operationId": "update_employee_shift_end_employee_shifts_shift_end_update__shift_id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shift Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeShiftUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeShiftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/shifts/": {
      "get": {
        "tags": [
          "shifts"
        ],
        "summary": "Get Shifts",
        "operationId": "get_shifts_shifts__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShiftOut"
                  },
                  "type": "array",
                  "title": "Response Get Shifts Shifts  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckListOut": {
        "properties": {
          "check_list": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Check List"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "check_list",
          "id"
        ],
        "title": "CheckListOut"
      },
      "CheckListUpdate": {
        "properties": {
          "check_list": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Check List"
          }
        },
        "type": "object",
        "required": [
          "check_list"
        ],
        "title": "CheckListUpdate"
      },
      "ClientCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ClientCreate"
      },
      "ClientOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "active"
        ],
        "title": "ClientOut"
      },
      "ClientUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ClientUpdate"
      },
      "EmployeeCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "position": {
            "$ref": "#/components/schemas/EmployeePosition"
          }
        },
        "type": "object",
        "required": [
          "name",
          "position"
        ],
        "title": "EmployeeCreate"
      },
      "EmployeeOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "position": {
            "$ref": "#/components/schemas/EmployeePosition"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "position",
          "id"
        ],
        "title": "EmployeeOut"
      },
      "EmployeePosition": {
        "type": "string",
        "enum": [
          "barista",
          "manager"
        ],
        "title": "EmployeePosition"
      },
      "EmployeeShiftCreate": {
        "properties": {
          "employee_id": {
            "type": "string",
            "format": "uuid",
            "title": "Employee Id"
          },
          "shift_id": {
            "title": "Shift Id",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object",
        "required": [
          "employee_id"
        ],
        "title": "EmployeeShiftCreate"
      },
      "EmployeeShiftOut": {
        "properties": {
          "employee_id": {
            "type": "string",
            "format": "uuid",
            "title": "Employee Id"
          },
          "shift_id": {
            "title": "Shift Id",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "title": "End Time",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object",
        "required": [
          "employee_id",
          "id",
          "active",
          "start_time"
        ],
        "title": "EmployeeShiftOut"
      },
      "EmployeeShiftUpdate": {
        "properties": {
          "last_employee_shift": {
            "type": "boolean",
            "title": "Last Employee Shift",
            "default": false
          }
        },
        "type": "object",
        "title": "EmployeeShiftUpdate"
      },
      "EmployeeShiftWithEmployeeOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "employee": {
            "$ref": "#/components/schemas/EmployeeOut"
          }
        },
        "type": "object",
        "required": [
          "id",
          "employee"
        ],
        "title": "EmployeeShiftWithEmployeeOut"
      },
      "EmployeeUpdate": {
        "properties": {
          "name": {
            "title": "Name",
            "type": [
              "string",
              "null"
            ]
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmployeePosition"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "EmployeeUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShiftOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "start_time": {
            "title": "Start Time",
            "type": [
              "string",
              "null"
            ]
          },
          "end_time": {
            "title": "End Time",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "active"
        ],
        "title": "ShiftOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}